+++ src/services/exportService.ts
@@ -42,61 +42,16 @@ import { updateFileCreationDateInEXIF } from './upload/exifService';
 import { MetadataObject } from 'types/upload';
 import QueueProcessor from './upload/queueProcessor';
 import { Collection } from 'types/collection';
-
-export type CollectionIDPathMap = Map<number, string>;
-export interface ExportProgress {
-    current: number;
-    total: number;
-}
-export interface ExportedCollectionPaths {
-    [collectionID: number]: string;
-}
-export interface ExportStats {
-    failed: number;
-    success: number;
-}
-
-const LATEST_EXPORT_VERSION = 1;
-
-export interface ExportRecord {
-    version?: number;
-    stage?: ExportStage;
-    lastAttemptTimestamp?: number;
-    progress?: ExportProgress;
-    queuedFiles?: string[];
-    exportedFiles?: string[];
-    failedFiles?: string[];
-    exportedCollectionPaths?: ExportedCollectionPaths;
-}
-export enum ExportStage {
-    INIT,
-    INPROGRESS,
-    PAUSED,
-    FINISHED,
-}
-
-enum ExportNotification {
-    START = 'export started',
-    IN_PROGRESS = 'export already in progress',
-    FINISH = 'export finished',
-    FAILED = 'export failed',
-    ABORT = 'export aborted',
-    PAUSE = 'export paused',
-    UP_TO_DATE = `no new files to export`,
-}
-
-enum RecordType {
-    SUCCESS = 'success',
-    FAILED = 'failed',
-}
-export enum ExportType {
-    NEW,
-    PENDING,
-    RETRY_FAILED,
-}
-
-const EXPORT_RECORD_FILE_NAME = 'export_status.json';
-export const METADATA_FOLDER_NAME = 'metadata';
+import {
+    ExportProgress,
+    ExportType,
+    ExportNotification,
+    CollectionIDPathMap,
+    RecordType,
+    ExportRecord,
+    EXPORT_RECORD_FILE_NAME,
+    LATEST_EXPORT_VERSION,
+} from 'types/export';
 
 class ExportService {
     ElectronAPIs: any;
